chore(coordinator): Move orderbook/routes.rs to routes module
