+++ coordinator/src/orderbook/mod.rs
@@ -1,7 +1,6 @@
 pub mod async_match;
 pub mod collaborative_revert;
 pub mod db;
-pub mod routes;
 pub mod trading;
 pub mod websocket;
